[MNG-7404] Easier way to mark one test method for a particular range of Maven versions
