+++ core-it-suite/src/test/java/org/apache/maven/it/MavenITmng2339BadProjectInterpolationTest.java
@@ -25,8 +25,6 @@ import java.io.File;
 
 /**
  * This is a test set for <a href="https://issues.apache.org/jira/browse/MNG-2339">MNG-2339</a>.
- * Part of this test has moved to {@link MavenITmng2339BadProjectInterpolationPre4Test} as it will no longer work
- * in Maven 4.
  */
 public class MavenITmng2339BadProjectInterpolationTest
     extends AbstractMavenIntegrationTestCase
@@ -51,6 +49,43 @@ public class MavenITmng2339BadProjectInterpolationTest
 
         verifier.verifyErrorFreeLog();
         verifier.resetStreams();
+    }
+
+    // test that -Dversion=1.0 is still available for interpolation.
+    public void testitMNG2339b()
+        throws Exception
+    {
+        requiresMavenVersion( "(2.0.8,4.0.0-alpha-1)" );
+        File testDir = ResourceExtractor.simpleExtractResources( getClass(), "/mng-2339/b" );
+
+        Verifier verifier;
+
+        verifier = newVerifier( testDir.getAbsolutePath() );
+        verifier.setAutoclean( false );
+        verifier.deleteDirectory( "target" );
+
+        verifier.setLogFileName( "log-pom-specified.txt" );
+        verifier.executeGoal( "initialize" );
+
+        assertTrue( "Touchfile using ${project.version} for ${version} does not exist.",
+                    new File( testDir, "target/touch-1.txt" ).exists() );
+
+        verifier.verifyErrorFreeLog();
+        verifier.resetStreams();
+
+        verifier = newVerifier( testDir.getAbsolutePath() );
+        verifier.setAutoclean( false );
+        verifier.deleteDirectory( "target" );
+
+        verifier.addCliOption( "-Dversion=2" );
+        verifier.setLogFileName( "log-cli-specified.txt" );
+        verifier.executeGoal( "initialize" );
+
+        verifier.verifyErrorFreeLog();
+        verifier.resetStreams();
+
+        assertTrue( "Touchfile using CLI-specified ${version} does not exist.",
+                    new File( testDir, "target/touch-2.txt" ).exists() );
     }
 
 }
